+++ data.js
@@ -1,0 +1,9 @@
+export const arrayOfStudents = [];
+
+export const houseArray = [
+  "Ravenclaw",
+  "Gryffindor",
+  "Hufflepuff",
+  "Slytherin",
+  "Expelled!",
+];
