+++ render-to-dom.js
@@ -1,0 +1,4 @@
+export const renderToDom = (divId, textToPrint) => {
+  const selectedDiv = document.querySelector(divId);
+  selectedDiv.innerHTML = textToPrint;
+};
